requirements.txt
Django~=2.2.4
pyTelegramBotAPI~=3.7.9
psycopg2-binary~=2.8.6
requests~=2.25.1
djangorestframework~=3.12.4